Move DesignSpace checks from Universal to UFO Sources profile
