+++ tests/profiles/ufo_sources_test.py
@@ -109,3 +109,49 @@ def test_check_unnecessary_fields(empty_ufo_font):
     font = TEST_FILE("source-sans-pro/OTF/SourceSansPro-Regular.otf")
     msg = assert_results_contain(check(font), SKIP, "unfulfilled-conditions")
     assert msg == "Unfulfilled Conditions: ufo_font"
+
+
+def test_check_designspace_has_sources():
+    """See if we can actually load the source files."""
+    check = CheckTester(ufo_sources_profile,
+                        "com.google.fonts/check/designspace_has_sources")
+
+    designspace = TEST_FILE("stupidfont/Stupid Font.designspace")
+    assert_PASS(check(designspace))
+
+    # TODO: FAIL, 'no-sources'
+
+
+def test_check_designspace_has_default_master():
+    """Ensure a default master is defined."""
+    check = CheckTester(ufo_sources_profile,
+                        "com.google.fonts/check/designspace_has_default_master")
+
+    designspace = TEST_FILE("stupidfont/Stupid Font.designspace")
+    assert_PASS(check(designspace))
+
+    # TODO: FAIL, 'not-found'
+
+
+def test_check_designspace_has_consistent_glyphset():
+    """Check consistency of glyphset in a designspace file."""
+    check = CheckTester(ufo_sources_profile,
+                        "com.google.fonts/check/designspace_has_consistent_glyphset")
+
+    designspace = TEST_FILE("stupidfont/Stupid Font.designspace")
+    assert_results_contain(check(designspace),
+                           FAIL, 'inconsistent-glyphset')
+
+    # TODO: Fix it and ensure it passes the check
+
+
+def test_check_designspace_has_consistent_codepoints():
+    """Check codepoints consistency in a designspace file."""
+    check = CheckTester(ufo_sources_profile,
+                        "com.google.fonts/check/designspace_has_consistent_codepoints")
+
+    designspace = TEST_FILE("stupidfont/Stupid Font.designspace")
+    assert_results_contain(check(designspace),
+                           FAIL, 'inconsistent-codepoints')
+
+    # TODO: Fix it and ensure it passes the check
